requirements.txt
fasttext_wheel==0.9.2
pandas==2.3.0
pymorphy2==0.9.1
Requests==2.32.3
torch==2.1.2
transformers==4.38.2
numpy==1.26.4
